Updated method CityClock.setLocalTimeByUtcZeroTime() to set the UTC zero time in data type long to the local time in data type Calendar.

+++ testfirst/src/main/java/tamingbadcode/testfirst/DeskClerk.java
@@ -22,7 +22,7 @@ public class DeskClerk {
         notifyAllCityClocks(this.cities.get(cityName).getUtcZeroTime(localTime));
     }
 
-    private void notifyAllCityClocks(int utcZeroTime) {
+    private void notifyAllCityClocks(long utcZeroTime) {
         for (String cityName : this.cities.keySet()) {
             this.cities.get(cityName).setLocalTimeByUtcZeroTime(utcZeroTime);
         }
